chore: updated app directory structure - updated directory structure - extracted routes - breakdown flows to pages and smaller components
