+++ package.json
@@ -32,7 +32,12 @@
       "react-app",
       "react-app/jest",
       "eslint:recommended"
-    ]
+    ],
+    "rules": {
+      "@typescript-eslint/no-unused-vars": [
+        "error"
+      ]
+    }
   },
   "browserslist": {
     "production": [
