+++ src/App.tsx
@@ -1,11 +1,14 @@
 import React from 'react'
 
-import { BrowserRouter as Router, Routes, Route, Link } from 'react-router-dom'
+import { BrowserRouter as Router, Routes, Route } from 'react-router-dom'
 
-import logo from './logo.svg'
 import './App.css'
 
-import BuyFlow, { ProductIds } from './buyFlow/Buyflow'
+import Header from './components/Header'
+import HomePage from './pages/HomePage'
+import BuyFlow from './pages/insurance/buyFlow/BuyFlow'
+import { ProductIds } from './pages/insurance/InsurancePage'
+import { ROUTES } from './routes/routes'
 
 
 const App: React.FC = () => {
@@ -13,18 +16,13 @@ const App: React.FC = () => {
   return (
     <Router>
       <div className="App">
-        <header className="App-header">
-          <img src={logo} className="App-logo" alt="logo" />
-        </header>
+        <Header />
         <Routes>
           <Route
-            path="/buy/insurance_dev"
+            path={ROUTES.INSURANCE.DEVELOPER_INSURANCE}
             element={<BuyFlow productId={ProductIds.devIns} />}
           />
-          <Route path="/" element={<>
-            <p>Welcome to Getsafe's Developer Insurance</p>
-            <Link to="/buy/insurance_dev">Get started!</Link>
-          </>} />
+          <Route path={ROUTES.HOME} element={<HomePage />} />
         </Routes>
       </div>
     </Router>
