+++ src/routes/routes.ts
@@ -1,0 +1,6 @@
+export const ROUTES = {
+  HOME: '/',
+  INSURANCE: {
+    DEVELOPER_INSURANCE: '/buy/insurance_dev',
+  },
+}
